scheduler: split up vm session management and guild handler
This is to allow for multiple vm sessions when using plugins

note: this is untested atm, it only compiles :^)
